+++ src/main/kotlin/blackjack/domain/model/PlayerInfo.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.domain.model
 
 class PlayerInfo(val name: String = "", val record: Record = Record())
