+++ src/main/kotlin/blackjack/domain/model/Round.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 class Round(private val trump: Trump = Trump(_cards = Trump.defaultTrump().shuffled().toMutableList())) {
     fun getCard(): Card = trump.getCard()
