+++ src/test/kotlin/blackjack/domain/model/RoundTest.kt
@@ -1,9 +1,8 @@
-package blackjack
+package blackjack.domain.model
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
-import java.lang.IllegalStateException
 
 class RoundTest {
