+++ src/test/kotlin/blackjack/domain/model/TrumpTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.model
 
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
