+++ src/test/kotlin/blackjack/util/InputParserTest.kt
@@ -1,5 +1,6 @@
-package blackjack
+package blackjack.util
 
+import blackjack.util.InputParser
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
